+++ examples/kafka_mirrormaker/mirrormaker.tf
@@ -53,4 +53,13 @@ resource "aiven_mirrormaker_replication_flow" "f1" {
     ".*\\.replica",
     "__.*"
   ]
+
+  config_properties_exclude = [
+    "follower\\.replication\\.throttled\\.replicas",
+    "leader\\.replication\\.throttled\\.replicas",
+    "message\\.timestamp\\.difference\\.max\\.ms",
+    "message\\.timestamp\\.type",
+    "unclean\\.leader\\.election\\.enable",
+    "min\\.insync\\.replicas"
+  ]
 }