+++ CHANGELOG.md
@@ -9,6 +9,10 @@ nav_order: 1
 <!-- Always keep the following header in place: -->
 <!-- [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
+## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
+- Add `config_properties_exclude` attribute to `aiven_mirrormaker_replication_flow` resource
+
 ## [4.19.0] - 2024-05-28
 
 - Add `is_super_admin` flag to `aiven_organization_application_user` resource